Add SEO Validation Tool
To speed up and partially automate SEO validation, a new tool is
added. This tool is a command line Python script.
Given a URL, it will crawl it, and for each page, do the following:
* check meta_title is non-empty and shorter than 60 chars
* check meta_description is non-empty and shorter than 160 chars
* check opengraph tags are non-empty
* check canonical_url is a valid url and corresponds to the page url

The script can provide its output on the command line or in a JSON
for further programmatic usage